refactored, putting components inside corresponding folder
